Added final features such as course signup handling, and report print outs.

+++ admin/del_user.php
@@ -4,8 +4,11 @@ if(!$session->isAdmin()){
 	header("Location: ../login.php");
 }
 $toDel = $_GET['u'];
-$q = "DELETE FROM users WHERE username='$toDel'";
-echo $q;
-$database->query($q);
+if($toDel != "admin") {
+	$q = "DELETE FROM users WHERE username='$toDel'";
+	$database->query($q);
+	$q = "DELETE FROM signups WHERE username='$toDel'";
+	$database->query($q);
+}
 header("Location: index.php?d=mu");
 ?>
